fixed lousy template filling
Basically what went wrong was we had a specific order in which end rooms and walls were sorted, so as to generate them in the correct direction. What I mean is, if the opening direction of a room is, for example, 1, the wall should match that direction and not be a wall for, lets say, the 3rd direction. So I'm scratching my head to see what changed in rGen or globalGen, when really its a lazy game scene edit of the rTemplate arrays. I know you didn't read all of that. Go back and contemplate what you've done.